Move `MainQueueDispatchDecorator` to it's own file
